refactor(cloudflare-metrics): split collector into resource-cache + emit + orchestration
the 736-line collector.ts was doing too much: REST resource lookups, row
→ metric translation, orchestration, and state persistence across isolate
invocations. break it up:

- resource-cache.ts    ResourceCacheService + module-level bulk caches
                       + resolveMissingZones; owns the per-tick cache and
                       the long-lived id→name maps
- emit.ts              pure functions: buildMetric, resolveTimestamp,
                       normalizeTagValue, applyResourceTags. no state, no
                       side effects — hands a ready-to-push Metric back
                       to the collector
- collector.ts         orchestration only: range/window math, batched
                       account fetches, zone iteration, scheduled-events
                       aggregation, per-dataset self-telemetry

collector.ts is 736 → 415 lines. the public api of CloudflareMetricsCollector
is unchanged, tests pass without modification, and
__resetCollectorCachesForTests is re-exported so existing imports keep
working.
